Initial deletion of code
