--- src/pqnstack/app/core/models.py
@@ -1,14 +1,0 @@
-import logging
-
-logger = logging.getLogger(__name__)
-
-
-def calculate_chsh_expectation_error(counts: list[int], dark_count: int = 0) -> float:
-    total_counts = sum(counts)
-    corrected_total = total_counts - 4 * dark_count
-    if corrected_total <= 0:
-        return 0
-    first_term = (total_counts**0.5) / corrected_total
-    expectation = abs(counts[0] + counts[3] - counts[1] - counts[2])
-    second_term = (expectation / corrected_total**2) * (total_counts + 4 * dark_count) ** 0.5
-    return float(first_term + second_term)
